+++ version.properties
@@ -2,4 +2,4 @@
 # it increments patch version based on most recent tag.
 # You can put explicit version here if needed, e.g. "1.0.0"
 # More information: https://github.com/shipkit/shipkit-auto-version/blob/master/README.md
-version=1.0.*
+version=2.0.*
